vm: implement Listable
